autokb/requirements.txt
aiohappyeyeballs==2.3.4 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.10.0 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anthropic==0.32.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.11" and python_version < "4.0"
arxiv==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.4.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0"
commonmark==0.9.1 ; python_version >= "3.11" and python_version < "4.0"
contourpy==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
dataclasses-json==0.6.7 ; python_version >= "3.11" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
duckduckgo-search==6.2.7 ; python_version >= "3.11" and python_version < "4.0"
exa-py==1.0.17 ; python_version >= "3.11" and python_version < "4.0"
feedparser==6.0.11 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.15.4 ; python_version >= "3.11" and python_version < "4.0"
fonttools==4.53.1 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
fsspec==2024.6.1 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.3 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
huggingface-hub==0.24.5 ; python_version >= "3.11" and python_version < "4.0"
idna==3.7 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.11" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.11" and python_version < "4.0"
langchain-anthropic==0.1.22 ; python_version >= "3.11" and python_version < "4.0"
langchain-community==0.2.10 ; python_version >= "3.11" and python_version < "4.0"
langchain-core==0.2.27 ; python_version >= "3.11" and python_version < "4.0"
langchain-exa==0.1.0 ; python_version >= "3.11" and python_version < "4.0"
langchain-openai==0.1.20 ; python_version >= "3.11" and python_version < "4.0"
langchain-text-splitters==0.2.2 ; python_version >= "3.11" and python_version < "4.0"
langchain==0.2.12 ; python_version >= "3.11" and python_version < "4.0"
langfuse==2.43.3 ; python_version >= "3.11" and python_version < "4.0"
langgraph-checkpoint==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
langgraph-cli==0.1.50 ; python_version >= "3.11" and python_version < "4.0"
langgraph==0.2.3 ; python_version >= "3.11" and python_version < "4.0"
langsmith==0.1.96 ; python_version >= "3.11" and python_version < "4.0"
loguru==0.5.3 ; python_version >= "3.11" and python_version < "4.0"
lxml==5.3.0 ; python_version >= "3.11" and python_version < "4.0"
marshmallow==3.21.3 ; python_version >= "3.11" and python_version < "4.0"
matplotlib==3.9.2 ; python_version >= "3.11" and python_version < "4.0"
mediawikiapi==1.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
nest-asyncio==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
networkx==3.3 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "4.0"
openai==1.38.0 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.10.6 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.11" and python_version < "4.0"
praw==7.7.1 ; python_version >= "3.11" and python_version < "4.0"
prawcore==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
primp==0.5.5 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytrends==4.9.2 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
regex==2024.7.24 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rich==10.16.2 ; python_version >= "3.11" and python_version < "4.0"
semanticscholar==0.8.4 ; python_version >= "3.11" and python_version < "4.0"
sgmllib3k==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.6 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.31 ; python_version >= "3.11" and python_version < "4.0"
stackapi==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
tavily-python==0.3.5 ; python_version >= "3.11" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.11" and python_version < "4.0"
tiktoken==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
tokenizers==0.19.1 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.4 ; python_version >= "3.11" and python_version < "4.0"
typer[all]==0.12.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0"
update-checker==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
websocket-client==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
wikibase-rest-api-client==0.1.4 ; python_version >= "3.11" and python_version < "4.0"
wikipedia==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
wrapt==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
